refactor: break table analyzer to smaller files
