+++ text-to-sql/google_spreadsheet_to_md/main.py
@@ -10,8 +10,9 @@ import logging
 from typing import Dict, Any
 from sheets import SheetProcessor
 from sheets.config import Config
-from csv_table_analyzer import TableDescriptionAnalyzer
+from table_analyzer import TableDescriptionAnalyzer
 from analysis_to_markdown import process_analysis_files
+from model_selector import ModelType, use_model
 
 # Configure logging
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
@@ -48,15 +49,29 @@ def download_spreadsheet(spreadsheet_id: str, output_dir: str) -> Dict[str, Any]
     logger.info("Processing Google Sheets data...")
     return process_spreadsheet(spreadsheet_id, output_dir=output_dir)
 
-def analyze_csv_files(output_dir: str) -> None:
+def analyze_csv_files(output_dir: str, use_llm: bool = True) -> None:
     """
     Analyze CSV files in the output directory.
     
     Args:
         output_dir: Directory containing CSV files to analyze
+        use_llm: Whether to use LLM-based analysis
     """
+    # Create chat model if LLM analysis is enabled
+    chat_model = None
+    if use_llm:
+        # Use DeepSeek as default
+        chat_model = use_model(
+            model_type=ModelType.DEEPSEEK,
+            temperature=0
+        )
+
     # Initialize and run CSV analyzer
-    analyzer = TableDescriptionAnalyzer(output_dir=output_dir)
+    analyzer = TableDescriptionAnalyzer(
+        output_dir=output_dir,
+        use_llm=use_llm,
+        chat_model=chat_model
+    )
     analyzer.process_all_files()
 
 def generate_markdown_docs(output_dir: str) -> None:
@@ -99,7 +114,7 @@ def main():
 
         # Step 2: Analyze generated CSV files
         logger.info("\nStep 2: Analyzing CSV files...")
-        analyze_csv_files(output_dir)
+        analyze_csv_files(output_dir, use_llm=not args.no_llm)
 
         # Step 3: Generate markdown documentation
         if not args.no_llm:
