+++ text-to-sql/google_spreadsheet_to_md/table_analyzer.py
@@ -1,0 +1,8 @@
+"""
+Main module for CSV table analysis.
+Exposes the FileProcessor class for analyzing CSV files.
+"""
+
+from csv_analysis.processors.file_processor import FileProcessor as TableDescriptionAnalyzer
+
+__all__ = ['TableDescriptionAnalyzer']
